Moving cache directory creation to the Source object instead of Path
