+++ src/runway/sources/source.py
@@ -7,6 +7,8 @@ Allows us to specify specific remote sourced resources for out application
 # pylint: disable=unused-import
 from typing import Dict, Optional, Union  # noqa: F401
 
+import os
+
 
 class Source(object):
     """
@@ -19,8 +21,13 @@ class Source(object):
     def __init__(self, config, cache_dir=None):
         # type(Dict[str, Union[str, Dict[str, str]]], Optional[str]) -> Source
         """Initialize."""
+
+        if not cache_dir:
+            cache_dir = os.path.expanduser("~/.runway_cache")  # type: str
+
         self.config = config  # type: Dict[str, Union[str, Dict[str, str]]]
         self.cache_dir = cache_dir  # type: str
+        self.__create_cache_directory()
 
     def fetch(self):
         # type: () -> None
@@ -34,3 +41,10 @@ class Source(object):
         for i in ['@', '/', ':']:
             uri = uri.replace(i, '_')  # type: str
         return uri
+
+    def __create_cache_directory(self):
+        # type: () -> None
+        """If no cache directory exists for the remote runway modules, create one."""
+        if not os.path.isdir(self.cache_dir):
+            os.mkdir(self.cache_dir)
+
